refactor(client/note): migrate individual note hooks to query module
1- Remove individual operation-specific hook files in favor of a more structured
query-based approach, separating data-fetching logic from component usage
following React Query best practices.

2- Create dedicated query and mutation files for all note operations (create, read,
update, delete) following React Query patterns for improved caching, error handling,
and server-state synchronization.
