+++ apps/client/src/features/note/queries/use-get-notes-by-video-id-query.ts
@@ -1,13 +1,14 @@
 "use client";
 
-import { useQuery } from "@tanstack/react-query";
-
 import type { IPaginationQueryDto } from "@tubenote/dtos";
 
+import { useQuery } from "@tanstack/react-query";
+
 import { useAuthStore } from "@/features/auth/store";
+
 import { getNotesByVideoId } from "../services";
 
-export function useGetNotesByVideoId({
+export function useGetNotesByVideoIdQuery({
   videoId,
   paginationQuery,
 }: { videoId: string; paginationQuery: IPaginationQueryDto }) {
@@ -17,7 +18,7 @@ export function useGetNotesByVideoId({
     queryKey: ["video-notes", videoId, paginationQuery],
     queryFn: () => getNotesByVideoId(videoId, paginationQuery),
     enabled: status === "authenticated",
-    select: (data) => ({
+    select: data => ({
       notes: data.payload.data,
       paginationMeta: data.payload.paginationMeta,
     }),
