+++ apps/client/src/features/note/queries/use-get-recent-notes-query.ts
@@ -6,13 +6,13 @@ import { useAuthStore } from "@/features/auth/store";
 
 import { getRecentNotes } from "../services";
 
-export function useGetRecentNotes() {
+export function useGetRecentNotesQuery() {
   const { status } = useAuthStore();
 
   return useQuery({
     queryKey: ["notes", "recent_notes"],
     queryFn: () => getRecentNotes(),
-    select: (data) => data.payload.data,
+    select: data => data.payload.data,
     enabled: status === "authenticated",
     staleTime: 1000 * 60 * 5, // 5 minutes
     refetchOnWindowFocus: false,
