+++ apps/client/src/features/note/queries/use-get-recently-updated-notes-query.ts
@@ -6,13 +6,13 @@ import { useAuthStore } from "@/features/auth/store";
 
 import { getRecentlyUpdatedNotes } from "../services";
 
-export function useGetRecentlyUpdatedNotes() {
+export function useGetRecentlyUpdatedNotesQuery() {
   const { status } = useAuthStore();
 
   return useQuery({
     queryKey: ["notes", "recently_updated_notes"],
     queryFn: () => getRecentlyUpdatedNotes(),
-    select: (data) => data.payload.data,
+    select: data => data.payload.data,
     enabled: status === "authenticated",
     staleTime: 1000 * 60 * 5, // 5 minutes
     refetchOnWindowFocus: false,
