+++ apps/client/src/features/note/queries/use-get-user-notes-query.ts
@@ -1,20 +1,20 @@
 "use client";
 
+import type { IPaginationQueryDto } from "@tubenote/dtos";
+
 import { useQuery } from "@tanstack/react-query";
 
 import { useAuthStore } from "@/features/auth/store";
 
-import type { IPaginationQueryDto } from "@tubenote/dtos";
-
 import { getUserNotes } from "../services";
 
-export function useGetUserNotes(paginationQuery: IPaginationQueryDto) {
+export function useGetUserNotesQuery(paginationQuery: IPaginationQueryDto) {
   const { status } = useAuthStore();
 
   return useQuery({
     queryKey: ["notes", paginationQuery],
     queryFn: () => getUserNotes(paginationQuery),
-    select: (data) => ({
+    select: data => ({
       notes: data.payload.data,
       paginationMeta: data.payload.paginationMeta,
     }),
